requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h5eee18b_5
ca-certificates=2023.12.12=h06a4308_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=39.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
ffmpeg=4.3=hf484d3e_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=py37h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
pandas=1.3.5=pypi_0
pillow=9.4.0=py37h6a678d5_0
pip=22.3.1=py37h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py37h06a4308_0
pysocks=1.7.1=py37_1
python=3.7.16=h7a1cb2a_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=1.12.1=py3.7_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=py37h06a4308_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py37_cu113
torchvision=0.13.1=py37_cu113
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
wheel=0.38.4=py37h06a4308_0
xz=5.4.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
